notebook reflects cleanup during meeting between Chen & John Thursday afternoon  - notebook now handles grammars with ocaml-style comments  - notebook includes oldstyle/newstyle comparison diagrams on the korean grammar
progress.txt describes how we analyzed the difference between sentence-initial NP-ACC and sentence-initial NP-NOM on this grammar. The NP-NOM cases could be the beginning of a very short intransitive sentence, and their intersection grammars are relatively low-entropy compared to the NP-ACC cases which necessarily are going to require more complex derivations, about which we know nothing. It thus makes a certain amount of sense that their intersection grammars would be higher entropy.

fixed the two .m files so that they do not include any error messages from FindRoot (about line-search not-working) in their output